--- specs/Naludrag/server/Protocol.java
@@ -1,12 +1,0 @@
-/**
- *
- * @author Olivier Liechti
- */
-public class Protocol {
-
-    public static final int PRESENCE_DEFAULT_PORT = 12302;
-
-    public static final String CMD_HELLO = "HELLO";
-    public static final String CMD_BYE = "BYE";
-
-}
